// src/types/Colors.ts
export enum Colors {
  Midnight = 'midnight',
  MidnightGreen = 'midnightgreen',
  Purple = 'purple',
  Red = 'red',
  Yellow = 'yellow',
  Pink = 'pink',
  Gold = 'gold',
  Green = 'green',
  White = 'white',
  Blue = 'blue',
  Silver = 'silver',
  Black = 'black',
  SpaceBlack = 'spaceblack',
  SpaceGray = 'spacegray',
  StarLight = 'starlight',
  Coral = 'coral',
  RoseGold = 'rose-gold',
  SkyBlue = 'sky-blue',
  Graphite = 'graphite',
  SierraBlue = 'sierrablue',
}